pyproject.toml
[project]
name = "naver-store-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-genai>=1.32.0",
    "gspread>=6.2.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "selenium>=4.35.0",
    "sentence-transformers>=5.1.0",
    "webdriver-manager>=4.0.2",
]

[dependency-groups]
dev = [
    "ruff==0.8.2",
]
